AI - refactor TerritoryManager to be able to handle attack and defense options simultaneously
